2026/WEB/mole-in-the-wall/web/requirements.txt
Flask
PyJWT
uvicorn
gunicorn
